drfcrud/task/serializers.py
from rest_framework import serializers
from .models import Task


from django.contrib.auth.models import User

class UserSerializer(serializers.ModelSerializer):
    tasks = serializers.PrimaryKeyRelatedField(many=True, queryset=Task.objects.all())
    
    class Meta:
        model = User
        fields = ['id', 'username', 'tasks']

class TaskSerializer(serializers.ModelSerializer):
    owner = serializers.ReadOnlyField(source='owner.username')
    class Meta:
        model = Task
        fields = ['id', 'taskName', 'description', 'dueDate', 'status', 'owner']